Release App 0.0.4

+++ scripts/init-extension.sh
@@ -1,15 +1,22 @@
 #!/usr/bin/env bash
 
-# Delete any existing common package in extension
-rm -rf packages/app/submodules/extension/.desktop
+if [ ! "$NO_COMMON_LINK" ]; then
+  # Use common workspace in extension
+  cd packages/app/submodules/extension
+  yarn link ../../../common
+  cd -
+fi
 
-# Copy common package to extension
-# Required as the extension is not a workspace so the common package is loaded using a local directory
-cp -r packages/common packages/app/submodules/extension/.desktop
+# Ensure an extension env file exists
+if [ ! -f "packages/app/submodules/extension/.metamaskrc" ]; then
+  if [ -f "packages/app/.metamaskrc" ]; then
+    cp packages/app/.metamaskrc packages/app/submodules/extension/
+  else
+    cp packages/app/.metamaskrc.dist packages/app/submodules/extension/.metamaskrc
+  fi
+fi
 
-# Copy the extension config from the app workspace into the extension
-if [ -f "packages/app/.metamaskrc" ]; then
-  cp packages/app/.metamaskrc packages/app/submodules/extension/
-else
-  echo ".metamaskrc file not found"
+# Ensure an app env file exists
+if [ ! -f "packages/app/.env" ]; then
+  cp packages/app/.env.example packages/app/.env
 fi
